Removed excess code leftover from migration
