+++ src/app/repositories/[...repoid]/Content.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Row, Col } from "src/components/Layout";
 
-import apolloClient from 'src/utils/server/apolloClient'
+import apolloClient from 'src/utils/apolloClient'
 import { Repository as RepositoryType } from 'src/data/types'
 import { REPOSITORY_QUERY, QueryData, QueryVar } from 'src/data/queries/repositoryQuery';
 
@@ -13,7 +13,7 @@ interface Props {
   isBot?: boolean
 }
 
-export default async function Content (props: Props) {
+export default async function Content(props: Props) {
   const { variables } = props
 
   const { data, error } = await apolloClient.query<QueryData, QueryVar>({
@@ -34,7 +34,7 @@ export default async function Content (props: Props) {
   return (
     <Row>
       <Col md={3}>
-        <RepositorySidebar repo={repository}/>
+        <RepositorySidebar repo={repository} />
       </Col>
       <Col md={9}>
         <RepositoryDetail repo={repository} />
