+++ src/app/ror.org/[rorid]/RelatedContent.tsx
@@ -8,7 +8,7 @@ import Loading from 'src/components/Loading/Loading'
 import { RELATED_CONTENT_QUERY, QueryVar, QueryData } from 'src/data/queries/organizationQuery'
 
 import Error from 'src/components/Error/Error'
-import WorksListing from 'src/components/WorksListing/Server'
+import WorksListing from 'src/components/WorksListing/WorksListing'
 import { pluralize } from 'src/utils/helpers';
 
 interface Props {
@@ -16,7 +16,7 @@ interface Props {
   isBot?: boolean
 }
 
-export default function RelatedContent (props: Props) {
+export default function RelatedContent(props: Props) {
   const { variables, isBot = false } = props
 
   const { loading, data, error } = useQuery<QueryData, QueryVar>(
@@ -49,7 +49,7 @@ export default function RelatedContent (props: Props) {
     : ''
 
   const totalCount = relatedWorks.totalCount
-  
+
   return (
     <>
       <Row>
