+++ src/components/AuthorsFacet/AuthorsFacet.tsx
@@ -1,14 +1,6 @@
 import React from 'react'
-import { Tooltip } from 'react-bootstrap'
-import OverlayTrigger from '../OverlayTrigger/OverlayTrigger'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import {
-    faSquare,
-    faCheckSquare,
-    faQuestionCircle
-} from '@fortawesome/free-regular-svg-icons'
-import { useRouter } from 'next/router'
-import Link from 'next/link'
+import FacetList from 'src/components/FacetList/FacetList'
+import { Facet } from 'src/data/types'
 
 type Props = {
     authors: Facet[]
@@ -17,45 +9,13 @@ type Props = {
     url: string
 }
 
-interface Facet {
-    id: string
-    title: string
-    count: number
-}
-
 const AuthorsFacet: React.FunctionComponent<Props> = ({
     authors,
     title,
     model,
     url
 }) => {
-    const router = useRouter()
-
-    function facetLink(param: string, value: string, id: string) {
-        let icon = faSquare
-
-        // get current query parameters from next router
-        const params = new URLSearchParams(router.query as any)
-
-        // delete model and cursor parameters
-        params.delete(model)
-        params.delete('cursor')
-
-        if (params.get(param) == value) {
-            // if param is present, delete from query and use checked icon
-            params.delete(param)
-            icon = faCheckSquare
-        } else {
-            // otherwise replace param with new value and use unchecked icon
-            params.set(param, value)
-        }
-
-        return (
-            <Link href={url + params.toString()} id={id}>
-                <FontAwesomeIcon icon={icon} />{' '}
-            </Link>
-        )
-    }
+    if (!authors || authors.length === 0) return null
 
     // Used for checking filter shouldnt show author that is already filtered
     function checkAuthorForPerson(author) {
@@ -75,43 +35,21 @@ const AuthorsFacet: React.FunctionComponent<Props> = ({
             return author
     }
 
-    return (
-        <React.Fragment>
-            {authors && authors.length > 0 && (
-                <div className="panel facets add">
-                    <div className="panel-body">
-                        <OverlayTrigger 
-                            placement="top"
-                            overlay={
-                                <Tooltip id="tooltipAuthors">
-                                    This list includes only {title} with ORCID iDs in DOI metadata.
-                                </Tooltip>
-                            }>
-                            <h4>{title} <FontAwesomeIcon icon={faQuestionCircle} /></h4>
-                        </OverlayTrigger>
-                        <ul id="authors-facets">
-                            {authors.filter(removeNullAuthors).filter(checkAuthorForPerson).map((facet) => (
-                                <li key={facet.id}>
-                                    {facetLink(
-                                        'filterQuery',
-                                        'creators_and_contributors.nameIdentifiers.nameIdentifier:"' +
-                                        facet.id +
-                                        '"',
-                                        'co-authors-facet-' + facet.id
-                                    )}
-                                    <div className="facet-title">{facet.title}</div>
-                                    <span className="number pull-right">
-                                        {facet.count.toLocaleString('en-US')}
-                                    </span>
-                                    <div className="clearfix" />
-                                </li>
-                            ))}
-                        </ul>
-                    </div>
-                </div>
-            )}
-        </React.Fragment>
-    )
+    function generateValue(facetId: string) {
+        return `creators_and_contributors.nameIdentifiers.nameIdentifier:"${facetId}"`
+    }
 
+    return (
+        <FacetList
+            data={authors.filter(removeNullAuthors).filter(checkAuthorForPerson)}
+            title={title}
+            id="authors-facets"
+            value={generateValue}
+            param="filterQuery"
+            url={url}
+            tooltipText={`This list includes only ${title} with ORCID iDs in DOI metadata.`}
+        />
+    )
 }
+
 export default AuthorsFacet
