+++ src/components/SearchRepository/SearchRepository.tsx
@@ -5,9 +5,9 @@ import Link from 'next/link'
 import { useQuery } from '@apollo/client'
 import { Row, Col, Alert } from 'react-bootstrap'
 
-import Pager from 'src/components/Pager/Server'
+import Pager from 'src/components/Pager/Pager'
 import FairFilter from 'src/components/FairFilter/FairFilter'
-import FacetList from 'src/components/FacetList/Server'
+import FacetList from 'src/components/FacetList/FacetList'
 import Error from 'src/components/Error/Error'
 import Loading from 'src/components/Loading/Loading'
 import RepositoryMetadata from 'src/components/RepositoryMetadata/RepositoryMetadata'
@@ -21,7 +21,7 @@ type Props = {
 }
 
 
-export default function SearchRepositories ({ variables }: Props) {
+export default function SearchRepositories({ variables }: Props) {
   const { loading, error, data } = useQuery<QueryData, QueryVar>(
     SEARCH_REPOSITORIES_GQL,
     {
@@ -46,22 +46,22 @@ export default function SearchRepositories ({ variables }: Props) {
   if (!repositories || repositories.nodes.length == 0) return (
     <Col md={9} mdOffset={3}>
       <div className="alert-works">
-      <Alert bsStyle="warning">
-        <p>No repositories found. If a domain repository is not available for your
-          kind of data, you may be able to use a general repository such as:</p>
+        <Alert bsStyle="warning">
+          <p>No repositories found. If a domain repository is not available for your
+            kind of data, you may be able to use a general repository such as:</p>
 
-        <ul>
-          <li><Link href="/repositories/10.17616/R34S33">Dryad</Link></li>
-          <li><Link href="/repositories/10.17616/R3PK5R">Figshare</Link></li>
-          <li><Link href="/repositories/10.17616/R3C880">Harvard Dataverse</Link></li>
-          <li><Link href="/repositories/10.17616/R3DD11">Mendeley Data</Link></li>
-          <li><Link href="/repositories/10.17616/R3N03T">Open Science Framework</Link></li>
-          <li><Link href="/repositories/10.17616/R3QP53">Zenodo</Link></li>
-        </ul>
+          <ul>
+            <li><Link href="/repositories/10.17616/R34S33">Dryad</Link></li>
+            <li><Link href="/repositories/10.17616/R3PK5R">Figshare</Link></li>
+            <li><Link href="/repositories/10.17616/R3C880">Harvard Dataverse</Link></li>
+            <li><Link href="/repositories/10.17616/R3DD11">Mendeley Data</Link></li>
+            <li><Link href="/repositories/10.17616/R3N03T">Open Science Framework</Link></li>
+            <li><Link href="/repositories/10.17616/R3QP53">Zenodo</Link></li>
+          </ul>
 
-        <p>You may also have an institutional repository or other local resources
-          at your organization available to you.  Contact your data librarian or
-          computing facility for local data services.</p>
+          <p>You may also have an institutional repository or other local resources
+            at your organization available to you.  Contact your data librarian or
+            computing facility for local data services.</p>
 
         </Alert>
       </div>
@@ -69,67 +69,67 @@ export default function SearchRepositories ({ variables }: Props) {
   )
 
 
-const renderFacets = () => {
-  if (repositories.totalCount == 0) return ""
+  const renderFacets = () => {
+    if (repositories.totalCount == 0) return ""
 
-  return (
-    <div className="col-md-3 hidden-xs hidden-sm" id="sidebar">
-      <div className="panel-group">
-        <FairFilter url="repositories/?" />
+    return (
+      <div className="col-md-3 hidden-xs hidden-sm" id="sidebar">
+        <div className="panel-group">
+          <FairFilter url="repositories/?" />
 
-        <FacetList
-          data={repositories.certificates || []}
-          title="Certificates"
-          id="certificate"
-          param="certificate"
-          url="repositories/?"
-        />
-        <FacetList
-          data={repositories.software || []}
-          title="Software"
-          id="software"
-          param="software"
-          url="repositories/?"
-        />
+          <FacetList
+            data={repositories.certificates || []}
+            title="Certificates"
+            id="certificate"
+            param="certificate"
+            url="repositories/?"
+          />
+          <FacetList
+            data={repositories.software || []}
+            title="Software"
+            id="software"
+            param="software"
+            url="repositories/?"
+          />
+        </div>
       </div>
-    </div>
-  )
-}
+    )
+  }
 
 
 
-const renderResults = () => {
-  const hasNextPage = repositories.pageInfo
-    ? repositories.pageInfo.hasNextPage
-    : false
+  const renderResults = () => {
+    const hasNextPage = repositories.pageInfo
+      ? repositories.pageInfo.hasNextPage
+      : false
 
-  const endCursor = repositories.pageInfo
-    ? repositories.pageInfo.endCursor
-    : ''
+    const endCursor = repositories.pageInfo
+      ? repositories.pageInfo.endCursor
+      : ''
 
 
-  return (
-    <Col md={9} id="content">
-      {(repositories.nodes.length || 0) > 0 && (
-        <h3 className="member-results">
-          {pluralize(repositories.totalCount || 0, 'Repository', false, 'Repositories')}
-        </h3>
-      )}
+    return (
+      <Col md={9} id="content">
+        {(repositories.nodes.length || 0) > 0 && (
+          <h3 className="member-results">
+            {pluralize(repositories.totalCount || 0, 'Repository', false, 'Repositories')}
+          </h3>
+        )}
 
-      {repositories.nodes.map((repo) => (
-        <RepositoryMetadata repo={repo} key={repo.id} />
-      ))}
-      
-      {(repositories.totalCount || 0) > 20 && (
-        <Pager
-          url={'/repositories?'}
-          hasNextPage={hasNextPage}
-          endCursor={endCursor}
-        />
-      )}
-    </Col>
-  )
-}
+        {repositories.nodes.map((repo) => (
+          <RepositoryMetadata repo={repo} key={repo.id} />
+        ))}
+
+        {(repositories.totalCount || 0) > 20 && (
+          <Pager
+            url={'/repositories?'}
+            hasNextPage={hasNextPage}
+            endCursor={endCursor}
+          />
+        )}
+      </Col>
+    )
+  }
 
   return (
     <Row>
